refactor(app): move networking tests
